# openapi.json
{
  "components": {
    "schemas": {
...
      "AnimalListOut": {
        "properties": {
          "animals": {
            "items": {
              "$ref": "#/components/schemas/AnimalListOut.AnimalOut"
            },
            "title": "Animals",
            "type": "array"
          }
        },
        "required": [
          "animals"
        ],
        "title": "AnimalListOut",
        "type": "object"
      },
...
      "CaptchaOut": {
        "description": "A registration captcha.\n\nThe digits are served as image URLs, not as text, which is the whole\npoint of it. The expected answer stays server-side in the session.",
        "properties": {
          "firstDigits": {
            "items": {
              "type": "string"
            },
            "title": "Firstdigits",
            "type": "array"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "secondDigits": {
            "items": {
              "type": "string"
            },
            "title": "Seconddigits",
            "type": "array"
          }
        },
        "required": [
          "operation",
          "firstDigits",
          "secondDigits"
        ],
        "title": "CaptchaOut",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorResponse.ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "OkOut": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkOut",
        "type": "object"
      },
...
      "SessionOut": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionOut.ActorOut"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          }
        },
        "required": [
          "authenticated"
        ],
        "title": "SessionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "description": "Model of a validation error response.",
        "items": {
          "$ref": "#/components/schemas/ValidationError.ValidationErrorElement"
        },
        "title": "ValidationErrorType",
        "type": "array"
      },
...
    }
  },
  "info": {
    "title": "kidallowance API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/allowances/{allowance_id}": {
      "delete": {
        "description": "The Jinja screen still removes an allowance with a GET, which means an <img src> can delete one. That route goes away at cutover.",
        "operationId": "delete__api_allowances_{allowance_id}",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "allowance_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "DELETE, not GET.",
        "tags": [
          "money"
        ]
      }
    },
    "/api/animals": {
      "get": {
        "description": "",
        "operationId": "get__api_animals",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalListOut"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "The animal alphabet a child login is built from.",
        "tags": [
          "kids"
        ]
      }
    },
    "/api/auth/captcha": {
      "get": {
        "description": "Kept in the session exactly as the Jinja flow does, so both registration paths validate against the same stored value.",
        "operationId": "get__api_auth_captcha",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptchaOut"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Issue a registration captcha and remember the answer.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/child/login": {
      "post": {
        "description": "",
        "operationId": "post__api_auth_child_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "child_login <POST>",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "",
        "operationId": "post__api_auth_logout",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "POST, not GET: logging out is a state change and must not be reachable from an <img> tag.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/parent/login": {
      "post": {
        "description": "",
        "operationId": "post__api_auth_parent_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParentLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "parent_login <POST>",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "description": "",
        "operationId": "get__api_auth_session",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Who is signed in. 200 even when nobody is, so the SPA can decide which of the three landing states to render without treating \"logged out\" as an error.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/csrf": {
      "get": {
        "description": "",
        "operationId": "get__api_csrf",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfOut"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "A CSRF token for a JSON client to send back as X-CSRFToken.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/dashboard": {
      "get": {
        "description": "",
        "operationId": "get__api_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "The home screen: a parent sees their kids, a child sees themselves.",
        "tags": [
          "kids"
        ]
      }
    },
...
    "/api/kids/available-animal-pairs": {
      "get": {
        "description": "",
        "operationId": "get__api_kids_available-animal-pairs",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalPairsOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Animal pairs still free for a nickname, for the registration screen.",
        "tags": [
          "kids"
        ]
      }
    },
...
  },
  "security": [],
...
}